Fix inspector fields work
