--- Assets/MixedRealityToolkit.SDK/Features/UX/Interactable/Scripts/Themes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2fba988d030ba9544b1df2ee210c6061
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
